requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.6
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
deepface==0.0.93
dill==0.3.9
Django==5.1.7
dlib==19.24.6
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==2.1.1
omegaconf==2.3.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.7
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
scipy==1.15.2
seaborn==0.13.2
setuptools==77.0.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
tf_keras==2.19.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.93
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
